refactor(tests): correct some imports
