+++ test/typed-dates.spec.ts
@@ -1,5 +1,5 @@
 import { TypedTestService } from './typed-test.service';
-import { TypedConfig } from '../src/lib/typed.config';
+import { TypedConfig } from '../src';
 
 describe('typed-dates', () => {
     let sampleService: TypedTestService;
