+++ test/typed-test.entities.ts
@@ -6,6 +6,12 @@ export class SomeClass {
     privateInfo: number;
 }
 
+export class SomeSimilarClass {
+    id: number;
+    text: string;
+    privateInfo: number;
+}
+
 export class SomeClassCopy {
     id: number;
     text: string;
